move duck interface to impl package
