+++ fabric-tool-attribute-api-v1/src/main/java/net/fabricmc/fabric/impl/tool/attribute/ItemStackContext.java
@@ -1,4 +1,4 @@
-package net.fabricmc.fabric.api.tool.attribute.v1;
+package net.fabricmc.fabric.impl.tool.attribute;
 
 import org.jetbrains.annotations.Nullable;
